Implement /NetworkDriver.CreateEndpoint
This patch replaces the mocked version of /NetworkDriver.CreateEndpoint
with the actual Neutron call. This unit test for the endpoint is also
implemented.

libnetwork's Endpoint is mapped into Neutron's Subnets and Port. The
request to create an Endpoint contains the information of libnetwork's
Interfaces to be manged, however in Neutron there's no single resource
corresponds to the Endpoint.

Therefore Kuryr tires to break the Endpoint into the following three
pieces.

1. Subnet for Address property in the request
   - Kuryr passes the given Address property to CIDR for the request
     against Neutron API as it is
2. Subnet for AddressIPv6 property in the request
   - Kuryr passes the given AddressIPv6 property to CIDR for the request
     against Neutron API as it is
3. Port which MAC address is the same as MacAddress property in the
   request

To distinguish Subnets created in 1. and 2., Kuryr names them by the
given EndpointId with the postfixes comprised of the indice of the
interfaces and the version of the IP addresses, i.e., "v4" or "v6". For
the Port, Kuryr gives the index and the prefix "port".

In the process 1. and 2., Kuryr doesn't specify the allocation list for
the addresses and it's the responsibility of Neutron to allocate the
appropriate set of the addresses.

In the latter workflow, where a container joins the network and be
assinged for the specific address, Neutron should pass the appropriate
IP addresses allocated in this phase.

Change-Id: I1798b3baff56e285059435f2b8620f36811b580f
Signed-off-by: Taku Fukushima <f.tac.mac@gmail.com>