Implement /NetworkDriver.DeleteEndpoint
This patch replaces the mocked version of /NetworkDriver.DeleteEndpoint
with the actual Neutron call. The unit test for the endpoint is also
implemented.

When the requst to delete the Endpoint is invoked, Kuryr deletes the
three resources corresponds to the Endpoint in the process of the
endpoint creation. Kuryrs searches the resources with the given
NetworkID and deletes them all based on the assumption there's no
conflict for the NetworkIDs.

The operation should be atomic but in this patch they're not and the
better design is required.

Change-Id: I4c94428c613094e758bd228f586cb73f507b5775
Signed-off-by: Taku Fukushima <f.tac.mac@gmail.com>